# docs/swagger/stadatitrasvsrv-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: API componente stacoresrv
  version: 1.0.0
  title: STADATITRASVSRV
  contact:
    email: andrea.fontanelli@eng.it
    
# tags are used for organizing operations
tags:
- name: test
  description: test resources
- name: comune
  description: Il comune
- name: nazione
  description: La nazione
- name: provincia
  description: La provincia
    
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
# basePath: /studiofontanelli/stacomunicazionesrv/1.0.0

schemes:
 - http

paths:
  /test/resources:
    get:
      tags:
        - test
      summary: Test resources
      description: Effettua il test dei servizi richiamati
      operationId: testResources
      produces:
      - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/GenericResponse'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
  
  /comune/{id}:
    get:  
      tags:
        - comune
      description: >
        Questa risorsa rappresenta il comune identificato con il suo ID.
      operationId: getComuneById
      produces:
      - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          format: int64
          description: > 
            la pk del comune 
      responses:
        200:
          schema:
            $ref: '#/definitions/Comune'
          description: OK
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Comune non trovato"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
            
  /comune:
    get:  
      tags:
        - comune
      description: >
        Questa risorsa rappresenta il comune identificato con il suo codice ISTAT.
      operationId: getComuneByIstat
      produces:
      - application/json
      parameters:
        - in: query
          required: true
          name: istat
          type: string
          description: > 
            il codice ISTAT del comune 
      responses:
        200:
          schema:
            $ref: '#/definitions/Comune'
          description: OK
        422:
          description: >  
            istat non valorizzato
        404:
          description: >
            Comune non trovato
        500:
          description: > 
            Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
  
  /comune/_search:
    post:  
      tags:
        - comune
      description: >
        Ricerca i comuni secondo i parametri di.
      operationId: findComune
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
        - in: body
          required: true
          name: request
          schema: 
            $ref: '#/definitions/SearchComuneRequest'
          description: > 
            I parametri di ricerca del comune
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Comune'
          description: OK
        422:
          description: >  
            request non valorizzata
        500:
          description: > 
            Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
  
  /province:
    get:  
      tags:
        - province
      description: >
        Ricerca le nazioni secondo i parametri di ricerca passati.
      operationId: findProvince
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
        
        - in: query
          name: search_type
          type: string
          enum: [all, piemonte]    
          description: >
            il tipo di ricerca da effettuare
             * all - tutte le province
             * piemonte - solo le province piemontesi 
            
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Provincia'
          description: OK
        422:
          description: >  
            request non valorizzata
        500:
          description: > 
            Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
            
  /nazione/{id}:
    get:  
      tags:
        - nazione
      description: >
        Questa risorsa rappresenta la nazione identificata con il suo ID.
      operationId: getNazioneById
      produces:
      - application/json
      parameters:
        - in: path
          required: true
          name: id
          type: integer
          format: int64
          description: > 
            la pk del comune 
      responses:
        200:
          schema:
            $ref: '#/definitions/Nazione'
          description: OK
        400:
          description: "Parametro non valorizzato"
        404:
          description: "Nazione non trovata"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   

  /nazione/_search:
    post:  
      tags:
        - nazione
      description: >
        Ricerca le nazioni secondo i parametri di ricerca passati.
      operationId: findNazione
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
        - in: body
          required: true
          name: request
          schema: 
            $ref: '#/definitions/SearchNazioneRequest'
          description: > 
            I parametri di ricerca della nazione
      responses:
        200:
          schema:
            type: array
            items:
              $ref: '#/definitions/Nazione'
          description: OK
        422:
          description: >  
            request non valorizzata
        500:
          description: > 
            Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'   
definitions:
  
  SearchComuneRequest:
    type: object
    description: Contiene i parametri di ricerca dei comuni 
    properties:
      denomonazione:
        type: string
        description: >
          il nome del comune
        example: 'TORINO'
      data_validita:
        type: string
        pattern: "^(3[01]|[12][0-9]|0[1-9])-(1[0-2]|0[1-9])-[0-9]{4}$"
        description: >
          la data di validità nel formato DD-MM-YYYY    
        example: 22-01-2022
      id_nazione:
        type: integer
        format: int64
        description: id nazione
 
  SearchNazioneRequest:
    type: object
    description: Contiene i parametri di ricerca delle nazioni 
    properties:
      denomonazione:
        type: string
        description: >
          il nome della nazione
        example: 'FRANCIA'
      data_validita:
        type: string
        pattern: "^(3[01]|[12][0-9]|0[1-9])-(1[0-2]|0[1-9])-[0-9]{4}$"
        description: >
          la data di validità nel formato DD-MM-YYYY    
        example: 15-08-2022
          
  GenericResponse:
    type: object
    description: Risposta di un servizio generico
    properties:
      message:
        type: string
        description: Contiene il messaggio eventuale del servizio chiamato
        example: 'Servizio ha risposto correttamente'
        
  ErrorDto:
    type: object
    description: Rappresenta l'errore restituito da servizio
    properties:
      status:
        type: integer
        description: status http restituito
        example: '500, 400, ecc'
      code:
        type: string
        description: code http restituito
        example: 'Internal server error, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
      error_details:
        type: array
        description: i dettagli errore
        items:
          $ref: '#/definitions/ErrorDetail'
  Comune:
    type: object
    description: Comune
    properties:
      id:
        type: integer
        format: int64
        description: pd comune 
      istat:
        type: string
        description: istat comune
      descrizione:
        type: string
        description: descrizione comune
      descrizione_estesa:
        type: string
        description: descrizione comune
      cap_riferimento:
        type: string
        description: cap riferimento
      provincia:
        type: string
        description: provincia
      codice_catastale:
        type: string
        description: codice catastale
  Nazione:
    type: object
    description: Nazione
    properties:
      id:
        type: integer
        format: int64
        description: id nazione
      descrizione:
        type: string
        description: >
          la descrizione della nazione
  Provincia:
    type: object
    description: La provincia
    properties:
      id:
        type: integer
        format: int64
        description: id provincia
      sigla:
        type: string
        description: >
          il codice della provincia
      istat:
        type: string
        description: >
          istat provincia
      descrizione:
        type: string
        description: >
          la descrizione della provincia
  ErrorDetail:
    type: object
    description: rappresenta il dettaglio errore restituito in caso di errori di validazione sui campi di input
    properties:
      field:
        type: string
        description: il nome del campo per cui è falita la validazione
        example: 'name, surname'
      error:
        type: string
        description: l'errore di validazione effettivo
        example: 'campo obbligatorio, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'